+++ v4/adw/adw-tab-view.go
@@ -889,7 +889,7 @@ func (x *TabPage) GetAtContext() *gtk.ATContext {
 // This functionality can be overridden by `GtkAccessible`
 // implementations, e.g. to get the bounds from an ignored
 // child widget.
-func (x *TabPage) GetBounds(XVar *int, YVar *int, WidthVar *int, HeightVar *int) bool {
+func (x *TabPage) GetBounds(XVar *int32, YVar *int32, WidthVar *int32, HeightVar *int32) bool {
 
 	cret := gtk.XGtkAccessibleGetBounds(x.GoPointer(), XVar, YVar, WidthVar, HeightVar)
 	return cret
@@ -1021,7 +1021,7 @@ func (x *TabPage) UpdateProperty(FirstPropertyVar gtk.AccessibleProperty, varArg
 // property change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *TabPage) UpdatePropertyValue(NPropertiesVar int, PropertiesVar []gtk.AccessibleProperty, ValuesVar []gobject.Value) {
+func (x *TabPage) UpdatePropertyValue(NPropertiesVar int32, PropertiesVar []gtk.AccessibleProperty, ValuesVar []gobject.Value) {
 
 	gtk.XGtkAccessibleUpdatePropertyValue(x.GoPointer(), NPropertiesVar, PropertiesVar, ValuesVar)
 
@@ -1057,7 +1057,7 @@ func (x *TabPage) UpdateRelation(FirstRelationVar gtk.AccessibleRelation, varArg
 // relation change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *TabPage) UpdateRelationValue(NRelationsVar int, RelationsVar []gtk.AccessibleRelation, ValuesVar []gobject.Value) {
+func (x *TabPage) UpdateRelationValue(NRelationsVar int32, RelationsVar []gtk.AccessibleRelation, ValuesVar []gobject.Value) {
 
 	gtk.XGtkAccessibleUpdateRelationValue(x.GoPointer(), NRelationsVar, RelationsVar, ValuesVar)
 
@@ -1094,7 +1094,7 @@ func (x *TabPage) UpdateState(FirstStateVar gtk.AccessibleState, varArgs ...inte
 // state change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *TabPage) UpdateStateValue(NStatesVar int, StatesVar []gtk.AccessibleState, ValuesVar []gobject.Value) {
+func (x *TabPage) UpdateStateValue(NStatesVar int32, StatesVar []gtk.AccessibleState, ValuesVar []gobject.Value) {
 
 	gtk.XGtkAccessibleUpdateStateValue(x.GoPointer(), NStatesVar, StatesVar, ValuesVar)
 
@@ -1366,30 +1366,30 @@ func (x *TabView) GetMenuModel() *gio.MenuModel {
 	return cls
 }
 
-var xTabViewGetNPages func(uintptr) int
+var xTabViewGetNPages func(uintptr) int32
 
 // Gets the number of pages in @self.
-func (x *TabView) GetNPages() int {
+func (x *TabView) GetNPages() int32 {
 
 	cret := xTabViewGetNPages(x.GoPointer())
 	return cret
 }
 
-var xTabViewGetNPinnedPages func(uintptr) int
+var xTabViewGetNPinnedPages func(uintptr) int32
 
 // Gets the number of pinned pages in @self.
 //
 // See [method@TabView.set_page_pinned].
-func (x *TabView) GetNPinnedPages() int {
+func (x *TabView) GetNPinnedPages() int32 {
 
 	cret := xTabViewGetNPinnedPages(x.GoPointer())
 	return cret
 }
 
-var xTabViewGetNthPage func(uintptr, int) uintptr
+var xTabViewGetNthPage func(uintptr, int32) uintptr
 
 // Gets the [class@TabPage] representing the child at @position.
-func (x *TabView) GetNthPage(PositionVar int) *TabPage {
+func (x *TabView) GetNthPage(PositionVar int32) *TabPage {
 	var cls *TabPage
 
 	cret := xTabViewGetNthPage(x.GoPointer(), PositionVar)
@@ -1420,10 +1420,10 @@ func (x *TabView) GetPage(ChildVar *gtk.Widget) *TabPage {
 	return cls
 }
 
-var xTabViewGetPagePosition func(uintptr, uintptr) int
+var xTabViewGetPagePosition func(uintptr, uintptr) int32
 
 // Finds the position of @page in @self, starting from 0.
-func (x *TabView) GetPagePosition(PageVar *TabPage) int {
+func (x *TabView) GetPagePosition(PageVar *TabPage) int32 {
 
 	cret := xTabViewGetPagePosition(x.GoPointer(), PageVar.GoPointer())
 	return cret
@@ -1475,13 +1475,13 @@ func (x *TabView) GetShortcuts() TabViewShortcuts {
 	return cret
 }
 
-var xTabViewInsert func(uintptr, uintptr, int) uintptr
+var xTabViewInsert func(uintptr, uintptr, int32) uintptr
 
 // Inserts a non-pinned page at @position.
 //
 // It's an error to try to insert a page before a pinned page, in that case
 // [method@TabView.insert_pinned] should be used instead.
-func (x *TabView) Insert(ChildVar *gtk.Widget, PositionVar int) *TabPage {
+func (x *TabView) Insert(ChildVar *gtk.Widget, PositionVar int32) *TabPage {
 	var cls *TabPage
 
 	cret := xTabViewInsert(x.GoPointer(), ChildVar.GoPointer(), PositionVar)
@@ -1495,13 +1495,13 @@ func (x *TabView) Insert(ChildVar *gtk.Widget, PositionVar int) *TabPage {
 	return cls
 }
 
-var xTabViewInsertPinned func(uintptr, uintptr, int) uintptr
+var xTabViewInsertPinned func(uintptr, uintptr, int32) uintptr
 
 // Inserts a pinned page at @position.
 //
 // It's an error to try to insert a pinned page after a non-pinned page, in
 // that case [method@TabView.insert] should be used instead.
-func (x *TabView) InsertPinned(ChildVar *gtk.Widget, PositionVar int) *TabPage {
+func (x *TabView) InsertPinned(ChildVar *gtk.Widget, PositionVar int32) *TabPage {
 	var cls *TabPage
 
 	cret := xTabViewInsertPinned(x.GoPointer(), ChildVar.GoPointer(), PositionVar)
@@ -1608,13 +1608,13 @@ func (x *TabView) ReorderLast(PageVar *TabPage) bool {
 	return cret
 }
 
-var xTabViewReorderPage func(uintptr, uintptr, int) bool
+var xTabViewReorderPage func(uintptr, uintptr, int32) bool
 
 // Reorders @page to @position.
 //
 // It's a programmer error to try to reorder a pinned page after a non-pinned
 // one, or a non-pinned page before a pinned one.
-func (x *TabView) ReorderPage(PageVar *TabPage, PositionVar int) bool {
+func (x *TabView) ReorderPage(PageVar *TabPage, PositionVar int32) bool {
 
 	cret := xTabViewReorderPage(x.GoPointer(), PageVar.GoPointer(), PositionVar)
 	return cret
@@ -1736,7 +1736,7 @@ func (x *TabView) SetShortcuts(ShortcutsVar TabViewShortcuts) {
 
 }
 
-var xTabViewTransferPage func(uintptr, uintptr, uintptr, int)
+var xTabViewTransferPage func(uintptr, uintptr, uintptr, int32)
 
 // Transfers @page from @self to @other_view.
 //
@@ -1744,7 +1744,7 @@ var xTabViewTransferPage func(uintptr, uintptr, uintptr, int)
 //
 // It's a programmer error to try to insert a pinned page after a non-pinned
 // one, or a non-pinned page before a pinned one.
-func (x *TabView) TransferPage(PageVar *TabPage, OtherViewVar *TabView, PositionVar int) {
+func (x *TabView) TransferPage(PageVar *TabPage, OtherViewVar *TabView, PositionVar int32) {
 
 	xTabViewTransferPage(x.GoPointer(), PageVar.GoPointer(), OtherViewVar.GoPointer(), PositionVar)
 
@@ -1777,20 +1777,20 @@ func (x *TabView) GetPropertyIsTransferringPage() bool {
 
 // GetPropertyNPages gets the "n-pages" property.
 // The number of pages in the tab view.
-func (x *TabView) GetPropertyNPages() int {
+func (x *TabView) GetPropertyNPages() int32 {
 	var v gobject.Value
 	x.GetProperty("n-pages", &v)
-	return v.GetInt()
+	return v.GetLong()
 }
 
 // GetPropertyNPinnedPages gets the "n-pinned-pages" property.
 // The number of pinned pages in the tab view.
 //
 // See [method@TabView.set_page_pinned].
-func (x *TabView) GetPropertyNPinnedPages() int {
+func (x *TabView) GetPropertyNPinnedPages() int32 {
 	var v gobject.Value
 	x.GetProperty("n-pinned-pages", &v)
-	return v.GetInt()
+	return v.GetLong()
 }
 
 // Emitted after [method@TabView.close_page] has been called for @page.
@@ -1897,13 +1897,13 @@ func (x *TabView) ConnectIndicatorActivated(cb *func(TabView, uintptr)) uint32 {
 //
 // A typical reason to connect to this signal would be to connect to page
 // signals for things such as updating window title.
-func (x *TabView) ConnectPageAttached(cb *func(TabView, uintptr, int)) uint32 {
+func (x *TabView) ConnectPageAttached(cb *func(TabView, uintptr, int32)) uint32 {
 	cbPtr := uintptr(unsafe.Pointer(cb))
 	if cbRefPtr, ok := glib.GetCallback(cbPtr); ok {
 		return gobject.SignalConnect(x.GoPointer(), "page-attached", cbRefPtr)
 	}
 
-	fcb := func(clsPtr uintptr, PageVarp uintptr, PositionVarp int) {
+	fcb := func(clsPtr uintptr, PageVarp uintptr, PositionVarp int32) {
 		fa := TabView{}
 		fa.Ptr = clsPtr
 		cbFn := *cb
@@ -1925,13 +1925,13 @@ func (x *TabView) ConnectPageAttached(cb *func(TabView, uintptr, int)) uint32 {
 // this function as the child might merely be moved to another window; use
 // child dispose handler for that or do it in sync with your
 // [method@TabView.close_page_finish] calls.
-func (x *TabView) ConnectPageDetached(cb *func(TabView, uintptr, int)) uint32 {
+func (x *TabView) ConnectPageDetached(cb *func(TabView, uintptr, int32)) uint32 {
 	cbPtr := uintptr(unsafe.Pointer(cb))
 	if cbRefPtr, ok := glib.GetCallback(cbPtr); ok {
 		return gobject.SignalConnect(x.GoPointer(), "page-detached", cbRefPtr)
 	}
 
-	fcb := func(clsPtr uintptr, PageVarp uintptr, PositionVarp int) {
+	fcb := func(clsPtr uintptr, PageVarp uintptr, PositionVarp int32) {
 		fa := TabView{}
 		fa.Ptr = clsPtr
 		cbFn := *cb
@@ -1945,13 +1945,13 @@ func (x *TabView) ConnectPageDetached(cb *func(TabView, uintptr, int)) uint32 {
 }
 
 // Emitted after @page has been reordered to @position.
-func (x *TabView) ConnectPageReordered(cb *func(TabView, uintptr, int)) uint32 {
+func (x *TabView) ConnectPageReordered(cb *func(TabView, uintptr, int32)) uint32 {
 	cbPtr := uintptr(unsafe.Pointer(cb))
 	if cbRefPtr, ok := glib.GetCallback(cbPtr); ok {
 		return gobject.SignalConnect(x.GoPointer(), "page-reordered", cbRefPtr)
 	}
 
-	fcb := func(clsPtr uintptr, PageVarp uintptr, PositionVarp int) {
+	fcb := func(clsPtr uintptr, PageVarp uintptr, PositionVarp int32) {
 		fa := TabView{}
 		fa.Ptr = clsPtr
 		cbFn := *cb
@@ -2046,7 +2046,7 @@ func (x *TabView) GetAtContext() *gtk.ATContext {
 // This functionality can be overridden by `GtkAccessible`
 // implementations, e.g. to get the bounds from an ignored
 // child widget.
-func (x *TabView) GetBounds(XVar *int, YVar *int, WidthVar *int, HeightVar *int) bool {
+func (x *TabView) GetBounds(XVar *int32, YVar *int32, WidthVar *int32, HeightVar *int32) bool {
 
 	cret := gtk.XGtkAccessibleGetBounds(x.GoPointer(), XVar, YVar, WidthVar, HeightVar)
 	return cret
@@ -2178,7 +2178,7 @@ func (x *TabView) UpdateProperty(FirstPropertyVar gtk.AccessibleProperty, varArg
 // property change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *TabView) UpdatePropertyValue(NPropertiesVar int, PropertiesVar []gtk.AccessibleProperty, ValuesVar []gobject.Value) {
+func (x *TabView) UpdatePropertyValue(NPropertiesVar int32, PropertiesVar []gtk.AccessibleProperty, ValuesVar []gobject.Value) {
 
 	gtk.XGtkAccessibleUpdatePropertyValue(x.GoPointer(), NPropertiesVar, PropertiesVar, ValuesVar)
 
@@ -2214,7 +2214,7 @@ func (x *TabView) UpdateRelation(FirstRelationVar gtk.AccessibleRelation, varArg
 // relation change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *TabView) UpdateRelationValue(NRelationsVar int, RelationsVar []gtk.AccessibleRelation, ValuesVar []gobject.Value) {
+func (x *TabView) UpdateRelationValue(NRelationsVar int32, RelationsVar []gtk.AccessibleRelation, ValuesVar []gobject.Value) {
 
 	gtk.XGtkAccessibleUpdateRelationValue(x.GoPointer(), NRelationsVar, RelationsVar, ValuesVar)
 
@@ -2251,7 +2251,7 @@ func (x *TabView) UpdateState(FirstStateVar gtk.AccessibleState, varArgs ...inte
 // state change must be communicated to assistive technologies.
 //
 // This function is meant to be used by language bindings.
-func (x *TabView) UpdateStateValue(NStatesVar int, StatesVar []gtk.AccessibleState, ValuesVar []gobject.Value) {
+func (x *TabView) UpdateStateValue(NStatesVar int32, StatesVar []gtk.AccessibleState, ValuesVar []gobject.Value) {
 
 	gtk.XGtkAccessibleUpdateStateValue(x.GoPointer(), NStatesVar, StatesVar, ValuesVar)
