+++ v4/gdk/gdkmonitor.go
@@ -122,10 +122,10 @@ func (x *Monitor) GetGeometry(GeometryVar *Rectangle) {
 
 }
 
-var xMonitorGetHeightMm func(uintptr) int
+var xMonitorGetHeightMm func(uintptr) int32
 
 // Gets the height in millimeters of the monitor.
-func (x *Monitor) GetHeightMm() int {
+func (x *Monitor) GetHeightMm() int32 {
 
 	cret := xMonitorGetHeightMm(x.GoPointer())
 	return cret
@@ -155,13 +155,13 @@ func (x *Monitor) GetModel() string {
 	return cret
 }
 
-var xMonitorGetRefreshRate func(uintptr) int
+var xMonitorGetRefreshRate func(uintptr) int32
 
 // Gets the refresh rate of the monitor, if available.
 //
 // The value is in milli-Hertz, so a refresh rate of 60Hz
 // is returned as 60000.
-func (x *Monitor) GetRefreshRate() int {
+func (x *Monitor) GetRefreshRate() int32 {
 
 	cret := xMonitorGetRefreshRate(x.GoPointer())
 	return cret
@@ -181,7 +181,7 @@ func (x *Monitor) GetScale() float64 {
 	return cret
 }
 
-var xMonitorGetScaleFactor func(uintptr) int
+var xMonitorGetScaleFactor func(uintptr) int32
 
 // Gets the internal scale factor that maps from monitor coordinates
 // to device pixels.
@@ -192,7 +192,7 @@ var xMonitorGetScaleFactor func(uintptr) int
 // This can be used if you want to create pixel based data for a
 // particular monitor, but most of the time you’re drawing to a surface
 // where it is better to use [method@Gdk.Surface.get_scale_factor] instead.
-func (x *Monitor) GetScaleFactor() int {
+func (x *Monitor) GetScaleFactor() int32 {
 
 	cret := xMonitorGetScaleFactor(x.GoPointer())
 	return cret
@@ -208,10 +208,10 @@ func (x *Monitor) GetSubpixelLayout() SubpixelLayout {
 	return cret
 }
 
-var xMonitorGetWidthMm func(uintptr) int
+var xMonitorGetWidthMm func(uintptr) int32
 
 // Gets the width in millimeters of the monitor.
-func (x *Monitor) GetWidthMm() int {
+func (x *Monitor) GetWidthMm() int32 {
 
 	cret := xMonitorGetWidthMm(x.GoPointer())
 	return cret
@@ -267,10 +267,10 @@ func (x *Monitor) GetPropertyGeometry() uintptr {
 
 // GetPropertyHeightMm gets the "height-mm" property.
 // The height of the monitor, in millimeters.
-func (x *Monitor) GetPropertyHeightMm() int {
+func (x *Monitor) GetPropertyHeightMm() int32 {
 	var v gobject.Value
 	x.GetProperty("height-mm", &v)
-	return v.GetInt()
+	return v.GetLong()
 }
 
 // GetPropertyManufacturer gets the "manufacturer" property.
@@ -291,10 +291,10 @@ func (x *Monitor) GetPropertyModel() string {
 
 // GetPropertyRefreshRate gets the "refresh-rate" property.
 // The refresh rate, in milli-Hertz.
-func (x *Monitor) GetPropertyRefreshRate() int {
+func (x *Monitor) GetPropertyRefreshRate() int32 {
 	var v gobject.Value
 	x.GetProperty("refresh-rate", &v)
-	return v.GetInt()
+	return v.GetLong()
 }
 
 // GetPropertyScale gets the "scale" property.
@@ -310,10 +310,10 @@ func (x *Monitor) GetPropertyScale() float64 {
 //
 // The scale factor is the next larger integer,
 // compared to [property@Gdk.Surface:scale].
-func (x *Monitor) GetPropertyScaleFactor() int {
+func (x *Monitor) GetPropertyScaleFactor() int32 {
 	var v gobject.Value
 	x.GetProperty("scale-factor", &v)
-	return v.GetInt()
+	return v.GetLong()
 }
 
 // GetPropertyValid gets the "valid" property.
@@ -326,10 +326,10 @@ func (x *Monitor) GetPropertyValid() bool {
 
 // GetPropertyWidthMm gets the "width-mm" property.
 // The width of the monitor, in millimeters.
-func (x *Monitor) GetPropertyWidthMm() int {
+func (x *Monitor) GetPropertyWidthMm() int32 {
 	var v gobject.Value
 	x.GetProperty("width-mm", &v)
-	return v.GetInt()
+	return v.GetLong()
 }
 
 // Emitted when the output represented by @monitor gets disconnected.
